requirements.txt
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==5.0.7
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.10
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
MarkupSafe==1.1.1
matplotlib-inline==0.1.2
numpy==1.20.2
pandas==1.2.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
ptyprocess==0.7.0
Pygments==2.9.0
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.15.0
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
